chore: rename quote to blockquote
